% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{create_initial_ratings}
\alias{create_initial_ratings}
\title{Create initial ratings}
\usage{
create_initial_ratings(player_groups, init_variance = .init_variance,
  group_diffs = NULL)
}
\description{
Create initial ratings
}
