% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{melt_match_results}
\alias{melt_match_results}
\title{Melt Match Results}
\usage{
melt_match_results(match_results)
}
\arguments{
\item{match_results}{}
}
\description{
Takes in match_results in winner/loser form
and turns it into name/opponent form
}
