% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{update_ratings}
\alias{update_ratings}
\title{Update Ratings}
\usage{
update_ratings(player_mean, player_variance, opponent_mean, opponent_variance,
  result)
}
\arguments{
\item{result}{}
}
\description{
Update Ratings
}
